bugfix: fixing a bugs of sprint #13 iter #2

+++ controllers/cards.js
@@ -6,7 +6,9 @@ module.exports.getCards = async (req, res) => {
 
     res.send(cards);
   } catch (err) {
+    console.log(`ERROR: ${err.name}`);
     console.log(`ERROR: ${err.message}`);
+
     res.status(500).send({ message: 'Ошибка на сервере' });
   }
 };
@@ -15,15 +17,23 @@ module.exports.getCard = async (req, res) => {
   const { _id } = req.params;
 
   try {
-    const cards = await Card.findById(_id);
+    const card = await Card.findById(_id);
 
-    res.send(cards);
-  } catch (err) {
-    if (err.name === 'CastError') {
+    if (!card) {
       res.status(404).send({ message: `Карточка с id: ${_id} не найдена!` });
+      return;
     }
 
+    res.send(card);
+  } catch (err) {
+    console.log(`ERROR: ${err.name}`);
     console.log(`ERROR: ${err.message}`);
+
+    if (err.name === 'CastError') {
+      res.status(404).send({ message: `Карточка с id: ${_id} не найдена!` });
+      return;
+    }
+
     res.status(500).send({ message: 'Ошибка на сервере' });
   }
 };
@@ -34,29 +44,40 @@ module.exports.createCard = async (req, res) => {
   try {
     const card = await Card.create({ name, link, owner: req.user._id });
 
-    card.validate();
     res.send(card);
   } catch (err) {
+    console.log(`ERROR: ${err.name}`);
+    console.log(`ERROR: ${err.message}`);
+
     if (err.name === 'ValidationError') {
       res.status(400).send({ message: 'Введены некорректные данные!' });
+      return;
     }
 
-    console.log(`ERROR: ${err.message}`);
     res.status(500).send({ message: 'Ошибка на сервере' });
   }
 };
 
 module.exports.deleteCard = async (req, res) => {
   try {
-    await Card.findByIdAndRemove(req.params._id);
+    const deletedCard = await Card.findById(req.params._id);
 
-    res.send({ message: 'Карточка удалена!' });
-  } catch (err) {
-    if (err.name === 'CastError') {
+    if (!deletedCard) {
       res.status(404).send({ message: `Карточка с id: ${req.params._id} не найдена!` });
+      return;
     }
 
+    await Card.findByIdAndRemove(req.params._id);
+    res.send({ message: 'Карточка удалена!' });
+  } catch (err) {
+    console.log(`ERROR: ${err.name}`);
     console.log(`ERROR: ${err.message}`);
+
+    if (err.name === 'CastError') {
+      res.status(404).send({ message: `Карточка с id: ${req.params._id} не найдена!` });
+      return;
+    }
+
     res.status(500).send({ message: 'Ошибка на сервере' });
   }
 };
